overlay: Fix positioning when the window position is not correct
Make the overlay positioning independent of global screen coordinates
when the overlay and the target widget are contained within the same
window.

When the window is resized (on OSX 10.11) PyQt4 only reports resize
events, even when the actual window position (top left corner of the window)
changes (e.g. by dragging the top window border).